# itdaing-web/openapi.json
{"openapi":"3.0.1","info":{"title":"Itdaing API","description":"Itdaing 팝업스토어 추천 서비스 API\n\n## 인증 방식\nJWT Bearer Token 기반 인증을 사용합니다.\n\n1. `/api/auth/login` 엔드포인트로 로그인하여 JWT 토큰을 발급받습니다.\n2. 발급받은 토큰을 Authorization 헤더에 포함하여 API를 호출합니다.\n   - 형식: `Authorization: Bearer {token}`\n\n## 사용자 역할\n- **CONSUMER**: 일반 소비자 (팝업스토어 검색, 위시리스트, 리뷰 작성)\n- **SELLER**: 판매자 (팝업스토어 등록 및 관리)\n- **ADMIN**: 관리자 (시스템 관리 및 승인)\n\n## JWT 토큰 구조\nJWT 토큰에는 다음 클레임이 포함됩니다:\n- `sub`: 사용자 ID\n- `role`: ROLE_CONSUMER, ROLE_SELLER, ROLE_ADMIN 중 하나\n- `iss`: itdaing-server\n- `iat`: 발급 시각 (Unix timestamp)\n- `exp`: 만료 시각 (발급 후 24시간)\n\n## 공개 API\n다음 엔드포인트는 인증 없이 접근 가능합니다:\n- `/api/auth/**`: 인증 관련 API (회원가입, 로그인)\n- `GET /api/master/**`: 마스터 데이터 조회 (지역, 스타일, 카테고리 등)\n","version":"v1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Master Data","description":"마스터 데이터 조회 API"},{"name":"Auth","description":"인증 API"}],"paths":{"/api/v1/auth/signup/seller":{"post":{"tags":["Auth"],"summary":"소비자 회원가입","description":"팝업스토어 판매자 계정을 생성합니다.\n\n가입된 사용자는 SELLER 역할이 부여되며, 다음 권한을 가집니다:\n- 팝업스토어 등록 및 관리\n- 판매자 프로필 관리\n- 팝업스토어 승인 요청\n- 고객 리뷰 조회\n\n로그인 시 JWT 토큰의 role 클레임에는 \"ROLE_SELLER\"가 설정됩니다.\n","operationId":"signupSeller","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupSellerRequest"}}},"required":true},"responses":{"409":{"description":"이메일 중복","content":{"application/json":{"example":{"success":false,"error":{"status":409,"code":"COMMON-409","message":"이미 사용 중인 이메일입니다"}}}}},"400":{"description":"입력값 검증 실패","content":{"application/json":{"example":{"success":false,"error":{"status":400,"code":"COMMON-001","message":"입력값이 올바르지 않습니다"}}}}},"200":{"description":"회원가입 성공","content":{"application/json":{"example":{"success":true,"data":{"userId":2,"email":"seller@example.com","role":"SELLER"}}}}}}}},"/api/v1/auth/signup/consumer":{"post":{"tags":["Auth"],"summary":"소비자 회원가입","description":"일반 소비자 계정을 생성합니다.\n\n가입된 사용자는 CONSUMER 역할이 부여되며, 다음 권한을 가집니다:\n- 팝업스토어 검색 및 조회\n- 위시리스트 관리\n- 리뷰 작성 및 조회\n- 소비자 선호 정보 설정\n\n로그인 시 JWT 토큰의 role 클레임에는 \"ROLE_CONSUMER\"가 설정됩니다.\n","operationId":"signupConsumer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupConsumerRequest"}}},"required":true},"responses":{"409":{"description":"이메일 중복","content":{"application/json":{"example":{"success":false,"error":{"status":409,"code":"COMMON-409","message":"이미 사용 중인 이메일입니다"}}}}},"400":{"description":"입력값 검증 실패","content":{"application/json":{"example":{"success":false,"error":{"status":400,"code":"COMMON-001","message":"입력값이 올바르지 않습니다"}}}}},"201":{"description":"회원가입 성공","content":{"application/json":{"example":{"success":true,"data":{"userId":1,"email":"consumer1@example.com","role":"CONSUMER"}}}}}}}},"/api/v1/auth/login":{"post":{"tags":["Auth"],"summary":"로그인","description":"이메일과 비밀번호로 로그인하여 JWT 액세스 토큰을 발급받습니다.\n\nJWT 토큰에는 다음 클레임이 포함됩니다:\n- sub: 사용자 ID\n- role: ROLE_CONSUMER, ROLE_SELLER, ROLE_ADMIN 중 하나\n- iss: itdaing-server\n- iat: 발급 시각 (Unix timestamp)\n- exp: 만료 시각 (발급 후 24시간)\n\n발급받은 토큰은 Authorization 헤더에 \"Bearer {token}\" 형식으로 포함하여 인증이 필요한 API를 호출할 수 있습니다.\n\n예시 토큰 페이로드 (디코딩 후):\n{\n  \"sub\": \"1\",\n  \"role\": \"ROLE_CONSUMER\",\n  \"iss\": \"itdaing-server\",\n  \"iat\": 1730419200,\n  \"exp\": 1730505600\n}\n","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"로그인 성공","content":{"application/json":{"examples":{"로그인 성공 응답":{"description":"소비자(CONSUMER) 계정으로 로그인한 경우","value":{"success":true,"data":{"accessToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwicm9sZSI6IlJPTEVfQ09OU1VNRVIiLCJpc3MiOiJpdGRhaW5nLXNlcnZlciIsImlhdCI6MTczMDQxOTIwMCwiZXhwIjoxNzMwNTA1NjAwfQ.signature"}}}}}}},"401":{"description":"인증 실패","content":{"application/json":{"example":{"success":false,"error":{"status":401,"code":"AUTH-401","message":"이메일 또는 비밀번호가 올바르지 않습니다"}}}}}}}},"/api/v1/users/me":{"get":{"tags":["Auth"],"summary":"내 프로필 조회","description":"인증된 사용자의 프로필 정보를 조회합니다.\n\n이 API는 JWT 토큰 인증이 필요합니다.\nAuthorization 헤더에 \"Bearer {token}\" 형식으로 토큰을 포함해야 합니다.\n\n응답의 role 필드는 다음 값 중 하나입니다:\n- CONSUMER: 일반 소비자\n- SELLER: 팝업스토어 판매자\n- ADMIN: 시스템 관리자\n","operationId":"getMyProfile","responses":{"200":{"description":"조회 성공","content":{"application/json":{"examples":{"소비자 프로필":{"description":"CONSUMER 역할의 사용자 프로필","value":{"success":true,"data":{"id":1,"email":"consumer@example.com","name":"김소비","nickname":"소비왕","role":"CONSUMER"}}},"판매자 프로필":{"description":"SELLER 역할의 사용자 프로필","value":{"success":true,"data":{"id":2,"email":"seller@example.com","name":"박판매","nickname":"팝업왕","role":"SELLER"}}}}}}},"404":{"description":"사용자를 찾을 수 없음","content":{"application/json":{"example":{"success":false,"error":{"status":404,"code":"COMMON-404","message":"사용자를 찾을 수 없습니다"}}}}},"401":{"description":"인증 실패 - 토큰이 없거나 유효하지 않음","content":{"application/json":{"example":{"success":false,"error":{"status":401,"code":"AUTH-401","message":"인증이 필요합니다"}}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/master/styles":{"get":{"tags":["Master Data"],"summary":"스타일 목록 조회","description":"팝업스토어 스타일 목록을 조회합니다","operationId":"getStyles","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListStyleResponse"}}}}}}},"/api/v1/master/regions":{"get":{"tags":["Master Data"],"summary":"지역 목록 조회","description":"광주광역시 구 단위 지역 목록을 조회합니다","operationId":"getRegions","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListRegionResponse"}}}}}}},"/api/v1/master/features":{"get":{"tags":["Master Data"],"summary":"특징 목록 조회","description":"팝업스토어 특징(편의사항) 목록을 조회합니다","operationId":"getFeatures","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListFeatureResponse"}}}}}}},"/api/v1/master/categories":{"get":{"tags":["Master Data"],"summary":"카테고리 목록 조회","description":"카테고리 목록을 조회합니다. type 파라미터로 POPUP 또는 CONSUMER 타입 필터링 가능","operationId":"getCategories","parameters":[{"name":"arg0","in":"query","description":"카테고리 타입 (POPUP|CONSUMER)","required":false,"schema":{"type":"string","enum":["POPUP","CONSUMER"]},"example":"POPUP"}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListCategoryResponse"}}}}}}}},"components":{"schemas":{"SignupSellerRequest":{"required":["email","loginId","name","password","passwordConfirm"],"type":"object","properties":{"email":{"type":"string","description":"이메일 주소","example":"seller@example.com"},"password":{"maxLength":20,"minLength":8,"type":"string","description":"비밀번호 (8-20자)","example":"password123"},"passwordConfirm":{"type":"string","description":"비밀번호 확인","example":"password123"},"loginId":{"maxLength":20,"minLength":4,"pattern":"^[a-z0-9_-]+$","type":"string","description":"로그인 ID(영문 소문자/숫자/하이픈/언더스코어, 4~20자)","example":"seller1"},"name":{"maxLength":100,"minLength":0,"type":"string","description":"이름","example":"박판매"},"nickname":{"maxLength":100,"minLength":0,"type":"string","description":"닉네임","example":"팝업왕"}},"description":"판매자 회원가입 요청"},"SignupConsumerRequest":{"required":["ageGroup","email","interestCategoryIds","loginId","name","password","passwordConfirm","regionIds","styleIds"],"type":"object","properties":{"email":{"type":"string","description":"이메일 주소","example":"consumer1@example.com"},"password":{"maxLength":20,"minLength":8,"type":"string","description":"비밀번호 (8-20자)","example":"P@ssw0rd1!"},"passwordConfirm":{"type":"string","description":"비밀번호 확인","example":"P@ssw0rd1!"},"loginId":{"maxLength":30,"minLength":4,"type":"string","description":"로그인 ID","example":"juchan01"},"name":{"maxLength":100,"minLength":0,"type":"string","description":"이름","example":"김소비"},"nickname":{"maxLength":100,"minLength":0,"type":"string","description":"닉네임","example":"소비왕"},"ageGroup":{"maximum":90,"minimum":10,"type":"integer","description":"나이대(10단위 정수)","format":"int32","example":20,"enum":[10,20,30,40,50,60,70,80,90]},"interestCategoryIds":{"maxItems":4,"minItems":1,"type":"array","description":"관심 카테고리 ID 목록(소비자용, 1~4개)","example":[101,105],"items":{"type":"integer","description":"관심 카테고리 ID 목록(소비자용, 1~4개)","format":"int64"}},"styleIds":{"maxItems":4,"minItems":1,"type":"array","description":"스타일 ID 목록(1~4개)","example":[12,15,19],"items":{"type":"integer","description":"스타일 ID 목록(1~4개)","format":"int64"}},"regionIds":{"maxItems":2,"minItems":1,"type":"array","description":"선호 지역 ID 목록(1~2개)","example":[2],"items":{"type":"integer","description":"선호 지역 ID 목록(1~2개)","format":"int64"}},"ageGroupTens":{"type":"boolean"}},"description":"소비자 회원가입 요청(기본정보 + 선호 선택을 한 번에 제출)"},"LoginRequest":{"required":["loginId","password"],"type":"object","properties":{"loginId":{"type":"string","description":"로그인 ID","example":"juchan01"},"password":{"type":"string","description":"비밀번호","example":"P@ssw0rd1"}},"description":"로그인 요청"},"ApiError":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"code":{"type":"string"},"message":{"type":"string"},"fieldErrors":{"type":"array","items":{"$ref":"#/components/schemas/FieldError"}}}},"ApiResponseListStyleResponse":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/StyleResponse"}},"error":{"$ref":"#/components/schemas/ApiError"}}},"FieldError":{"type":"object","properties":{"field":{"type":"string"},"value":{"type":"string"},"reason":{"type":"string"}}},"StyleResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"ApiResponseListRegionResponse":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/RegionResponse"}},"error":{"$ref":"#/components/schemas/ApiError"}}},"RegionResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"ApiResponseListFeatureResponse":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/FeatureResponse"}},"error":{"$ref":"#/components/schemas/ApiError"}}},"FeatureResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"ApiResponseListCategoryResponse":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/CategoryResponse"}},"error":{"$ref":"#/components/schemas/ApiError"}}},"CategoryResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"type":{"type":"string","enum":["POPUP","CONSUMER"]}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT 액세스 토큰을 입력하세요.\n\n토큰 발급:\n1. POST /api/auth/login으로 로그인\n2. 응답의 accessToken 값을 복사\n3. 우측 상단 'Authorize' 버튼 클릭 후 토큰 입력\n\n예시: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwicm9sZSI6IlJPTEVfQ09OU1VNRVIifQ...\n","scheme":"bearer","bearerFormat":"JWT"}}}}